diet(phase9): split engine/call_graph.rs (1894 LOC) into 9-file module
Mechanical split, behavior unchanged. Public API (lib.rs:60 re-exports)
preserved via mod.rs `pub use`. queries.rs submodule (PR #291) untouched.

call_graph/
- mod.rs (17)         re-exports
- types.rs (41)       CallEdge, CallerEntry, CalleeEntry
- noise.rs (59)       is_noise_callee, is_noise_callee_for_lang
- language.rs (89)    CallLanguageConfig + path/lang detection helpers
- js_imports.rs (148) JSImportBinding + JSImportBindingIndex +
                      parse/build/filter/insert helpers
- extract.rs (236)    cached_call_query, extract_calls,
                      extract_calls_from_source + collect helpers
- resolve.rs (219)    resolve_call_edges, maybe_import_graph,
                      collect_candidate_files
- api.rs (146)        get_callers, get_callees
- tests.rs (983)      `mod tests` content moved from inline
- queries.rs (171)    UNCHANGED (PR #291 submodule)

Boundary checks confirmed:
- lib.rs:60 `pub use call_graph::{CallEdge, CalleeEntry, CallerEntry,
  extract_calls, get_callees, get_callers}` resolves unchanged
- scip_backend.rs:16, import_graph/dead_code.rs:1,
  symbols/writer.rs:38,290 all unchanged

Gates:
- cargo check --workspace --features http
- cargo test -p codelens-engine (all passed)
- cargo test -p codelens-mcp --bin codelens-mcp (593 passed in isolation;
  prepare_harness_session_warns_when_daemon_binary_is_stale shows the
  documented timing flake under parallel load — see test comment, PRs
  #174/#177/#184/#185/#187 — passes solo)
- cargo fmt --all -- --check
